feat: add dashboard and cart with async storage
This includes the functionalities required to add products to the cart,
increment and decrement the quantity of products directly from the cart,
calculations on the total of items in cart and total price of the
purchase. All of this including the use of Async Storage to save the
cart and reload it when reopening the application.